requirements.txt


pillow
Flask
Flask-SQLAlchemy
Celery
redis
Pillow
requests